+++ simpledbus/include/simpledbus/base/Connection.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <dbus/dbus.h>
+#include <condition_variable>
 #include <mutex>
 #include <unordered_map>
 #include <functional>
@@ -27,6 +28,7 @@ class Connection {
     Message pop_message();
 
     void send(Message& msg);
+    Message send_with_reply(Message& msg);
     Message send_with_reply_and_block(Message& msg);
 
     bool register_object_path(const std::string& path, std::function<void(Message&)> handler);
@@ -42,9 +44,17 @@ class Connection {
     ::DBusConnection* _conn;
 
     std::recursive_mutex _mutex;
+    std::unordered_map<std::string, std::function<void(Message&)>> _message_handlers;
 
     static DBusHandlerResult static_message_handler(DBusConnection* connection, DBusMessage* message, void* user_data);
-    std::unordered_map<std::string, std::function<void(Message&)>> _message_handlers;
+    static void static_reply_handler(DBusPendingCall* pending, void* user_data);
+
+    struct AsyncContext {
+        DBusMessage* reply = nullptr;
+        bool completed = false;
+        std::mutex mtx;
+        std::condition_variable cv;
+    };
 };
 
 }  // namespace SimpleDBus